MISUV-3860   * minor code refactoring;
